requirements.txt
python-telegram-bot==21.9
groq==0.13.1
python-dotenv==1.0.1
fastapi==0.115.4
uvicorn[standard]==0.32.1